requirements.txt
keyboard
requests
python-dateutil